God I need to get a life.
1st and most important thing is this....

Go to these 2 locations and remove the teleporters in both felucca and trammel.

588 1637 0.  Remove the teleporter you are standing on.

6575 889 0.  Remove both teleporters you are standing at.

Once all 6 teleporters (3 trammel, 3 felucca) are deleted and the rest of the updates are installed.  Either re decorate with admin menu or use the admin menu to remove all decoration and then re decorate.

the first teleporter will allow access to the Blighted Grove without the Bone Machete.  The other 2 will take you outside the tree.

Once you redecorate the 2 teelporters inside the tree will reappear but not take you outside the tree.  Instead it will take you to another set of teleporters that take you to that spot.

Fixed the Quell's hue.  Modified Meraktus and the Tormented Minotaur.  Updated the Talisman Slayer to work with the Tormented Minotaurs.  Added 3 animated rocks with water tiles for the artifact system.  Regular water tiles will no longer be dropped.