Updated to RunUO SVN 649.
Plus a little code cleanup.

A new exe will be required.  I also included a new Ultima.DLL.  This should fix a few gump issues in Mondain's.  Please download the new Ultima.DLL when compiling your new exe or downloading a precompiled exe.

Previous Ultima.dll is still available as a depreciated download in the download section of the SVN.